+++ changelog.md
@@ -5,6 +5,8 @@
 
 ### changed sttuff
 * add `None()`, deprecate `Py_None()`
+* change `@cimport` into `b.addTranslateC` and `c` module, make `py` module able to use debug build.
+* `py` module no longer links libc.
 
 ### breaking
 * Change `PyCFunction` definition `?*const fn ([*c]c.PyObject, [*c]c.PyObject) callconv(.c) ?*c.PyObject` ->